+++ lib/modbot/modbot_check.rb
@@ -71,7 +71,7 @@ module ModbotCheck
   def test_condition(condition, item, test_item)
     test = query_test_condition(condition, test_item)
     if test.kind_of?(Integer) || test == true
-      item.verdict << [condition.action, condition.weight]
+      item.verdict.unshift([condition.action, condition.weight])
       test_result = true
       @l.info "#{test_item} ::: #{condition.query} #{condition.what} ::: #{test_result}, recommend #{condition.action}"
     elsif test.nil? || test == false